pyproject.toml
[tool.poetry]
name = "eo-data-pipeline"
version = "0.1.0"
description = "Pipeline for fetching data from Earth Search Catalog and dumping it to a local databse for further downstream applications like Analytics"
authors = ["Akhil Singh Rana <akhilsinghrana@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.11,<3.13"
xarray = "^2024.6.0"
netCDF4 = "^1.7.1.post1"
h5netcdf = "^1.3.0"
apache-airflow = "^2.9.3"

# dev dependencies
pystac-client = "^0.8.2"
hydra-core = "^1.3.2"
dask = "^2024.7.0"
omegaconf = "^2.3.0"
prefect = "^2.19.8"
planetary-computer = "^1.0.0"
shapely = "^2.0.5"
[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
black = "^23.3.0"
pylint = "^3.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
